requirements.txt
absl-py==2.1.0
asgiref==3.8.1
asttokens==2.4.1
astunparse==1.6.3
attrs==23.2.0
blinker==1.8.2
capstone==5.0.3
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comm==0.2.2
comtypes==1.4.6
contourpy==1.2.1
cycler==0.12.1
debugpy==1.8.2
decorator==5.1.1
distlib==0.3.8
Django==5.1
docutils==0.21.2
executing==2.0.1
Flask==3.0.3
flatbuffers==24.3.25
fonttools==4.53.1
gast==0.6.0
google-pasta==0.2.0
h11==0.14.0
idna==3.7
ipykernel==6.29.5
ipython==8.26.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
joblib==1.4.2
jupyter_client==8.6.2
jupyter_core==5.7.2
keras==3.4.1
kiwisolver==1.4.5
libclang==18.1.1
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.1
matplotlib-inline==0.1.7
mdurl==0.1.2
ml-dtypes==0.4.0
namex==0.0.8
nest-asyncio==1.6.0
nltk==3.8.1
numpy==1.26.4
opt-einsum==3.3.0
optree==0.12.1
outcome==1.3.0.post0
packaging==24.1
pandas==2.2.2
parso==0.8.4
pefile==2023.2.7
pillow==10.4.0
platformdirs==4.2.2
prompt_toolkit==3.0.47
protobuf==4.25.4
pure_eval==0.2.3
pycparser==2.22
Pygments==2.18.0
pyparsing==3.1.2
pypiwin32==223
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
pywin32==306
pywin32-ctypes==0.2.3
pyzmq==26.0.3
regex==2024.7.24
requests==2.32.3
scikit-learn==1.5.1
scipy==1.14.0
setuptools==72.1.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
sqlparse==0.5.1
stack-data==0.6.3
tensorboard==2.17.0
tensorboard-data-server==0.7.2
tensorflow==2.17.0
tensorflow-intel==2.17.0
termcolor==2.4.0
threadpoolctl==3.5.0
tornado==6.4.1
tqdm==4.66.4
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
wcwidth==0.2.13
websocket-client==1.8.0
Werkzeug==3.0.3
wheel==0.43.0
wrapt==1.16.0
wsproto==1.2.0